# Conflicts:
#	hparams.py
#	sentences.txt

+++ utils/dataset.py
@@ -215,7 +215,7 @@ def collate_tts(batch, r):
     if len(batch[0]) > 4:
         dur = [pad1d(x[4][:max_x_len], max_x_len) for x in batch]
         dur = np.stack(dur)
-        dur = torch.tensor(dur)
+        dur = torch.tensor(dur).float()
         return chars, mel, ids, mel_lens, dur
     else:
         return chars, mel, ids, mel_lens
